Fix: Observation space dtype mismatch (float64 → float32)
All 4 state interpreters defined observation_space as float64, but
actual observations are converted to float32 at state.py:108.

This mismatch caused Gymnasium warnings:
  - 'obs dtype expecting float32, got float64'
  - 'obs not within observation space'

Fix: Change all get_observation_space() dtypes to float32 to match
the actual observation array dtype.

+++ SplunkResearch/src/wrappers/state_interpreters.py
@@ -79,7 +79,7 @@ class SparseStepInterpreter(StateInterpreter):
         return spaces.Box(
             low=0, high=1,
             shape=(n_logtypes * 2 + total_steps,),
-            dtype=np.float64,
+            dtype=np.float32,
         )
 
     def normalize(self, state):
@@ -110,7 +110,7 @@ class KLDivergenceInterpreter(StateInterpreter):
         return spaces.Box(
             low=0, high=3,
             shape=(n_logtypes * 2 + 3,),
-            dtype=np.float64,
+            dtype=np.float32,
         )
 
     def normalize(self, state):
@@ -152,7 +152,7 @@ class AccumulatedLogVolumeInterpreter(StateInterpreter):
         return spaces.Box(
             low=0, high=1,
             shape=(n_logtypes * 2 + extra,),
-            dtype=np.float64,
+            dtype=np.float32,
         )
 
     def normalize(self, state):
@@ -198,7 +198,7 @@ class AlertAwareInterpreter(AccumulatedLogVolumeInterpreter):
         return spaces.Box(
             low=0, high=1,
             shape=(n_logtypes * 2 + base_extra + alert_extra,),
-            dtype=np.float64,
+            dtype=np.float32,
         )
 
     def build_state(self, ctx: StateContext) -> np.ndarray:
